requirements.txt
azure-functions
requests
azure-storage-blob
azure-keyvault-secrets
azure-identity